eoscompanion: Split out routes code into middlewares and v1 routes
We'll leverage this structure in a later commit to create v2 routes
