+++ Makefile.am.service.inc
@@ -31,6 +31,7 @@ app_PYTHON += \
 	eoscompanion/applications_query.py \
 	eoscompanion/constants.py \
 	eoscompanion/content_streaming.py \
+	eoscompanion/core_routes.py \
 	eoscompanion/dummy_feed.py \
 	eoscompanion/eknservices_bridge.py \
 	eoscompanion/ekn_content_adjuster.py \
@@ -40,9 +41,12 @@ app_PYTHON += \
 	eoscompanion/functional.py \
 	eoscompanion/license_content_adjuster.py \
 	eoscompanion/main.py \
+	eoscompanion/middlewares.py \
 	eoscompanion/responses.py \
 	eoscompanion/routes.py \
+	eoscompanion/server.py \
 	eoscompanion/service.py \
+	eoscompanion/v1_routes.py \
 	$(NULL)
 
 # Wrapper script
