+++ eoscompanion/server.py
@@ -1,0 +1,45 @@
+# /eoscompanion/server.py
+#
+# Copyright (C) 2017, 2018 Endless Mobile, Inc.
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, write to the Free Software Foundation, Inc.,
+# 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+# All rights reserved.
+'''Server constructor func for eoscompanion.'''
+
+from gi.repository import Soup
+
+from .middlewares import (
+    application_hold_middleware,
+    cancellability_middleware,
+    handle_404_middleware
+)
+from .routes import create_companion_app_routes
+
+def create_companion_app_webserver(application, content_db_conn):
+    '''Create a HTTP server with companion app routes.'''
+    server = Soup.Server()
+    for path, handler in create_companion_app_routes(content_db_conn).items():
+        server.add_handler(
+            path,
+            cancellability_middleware(
+                handle_404_middleware(
+                    path,
+                    application_hold_middleware(application,
+                                                handler)
+                )
+            )
+        )
+
+    return server
